fix: Text -> Typo
