+++ packages/react/src/components/Typo/Detail.tsx
