+++ packages/react/src/components/Typo/Display.tsx
