+++ packages/react/src/components/Typo/__Links__.stories.tsx
@@ -1,13 +1,13 @@
 import {
-  Label,
-  type LabelProps
-} from './Label';
+  Links,
+  type LinksProps
+} from './Links';
 
-const sizes = ['lg','md','sm','xs'] as const;
+const sizes = ['lg','md','sm'] as const;
 
 export default {
-  Label: 'Components/Text/Label',
-  component: Label,
+  Links: 'Components/Typo/Links',
+  component: Links,
   parameters: {
     layout: 'centered',
   },
@@ -15,11 +15,11 @@ export default {
   argTypes: {
     size: {
       options: sizes,
-      description: 'Label 폰트 크기',
+      description: 'Links 폰트 크기',
       control: { type: 'select' }
     },
     bold: {
-      description: 'Label 폰트 굵기',
+      description: 'Links 폰트 굵기',
       control: { type: 'boolean'}
     }
   },
@@ -30,16 +30,16 @@ export default {
 };
 
 export const Default = {
-  render: (props: LabelProps<'p'>) => {
+  render: (props: LinksProps<'p'>) => {
     return (
       <section className="p-4 w-480px">
-        <Label as="p" {...props}>
+        <Links as="p" {...props}>
           다람쥐 헌 쳇바퀴에 타고파
-        </Label>
+        </Links>
         <br/>
-        <Label as="p" {...props}>
+        <Links as="p" {...props}>
           The quick brown fox jumps over the lazy dog
-        </Label>
+        </Links>
       </section>
     );
   },
@@ -49,19 +49,19 @@ export const Bold = {
   render: () => {
     return (
       <section className="p-4 w-480px">
-        <Label as="p">
+        <Links as="p">
           다람쥐 헌 쳇바퀴에 타고파
-        </Label>
-        <Label as="p" bold>
+        </Links>
+        <Links as="p" bold>
           다람쥐 헌 쳇바퀴에 타고파
-        </Label>
+        </Links>
         <br/>
-        <Label as="p">
+        <Links as="p">
           The quick brown fox jumps over the lazy dog
-        </Label>
-        <Label as="p" bold>
+        </Links>
+        <Links as="p" bold>
           The quick brown fox jumps over the lazy dog
-        </Label>
+        </Links>
       </section>
     );
   },
@@ -75,9 +75,9 @@ export const Size = {
         <section className='flex flex-col gap-4'>
         {
           sizes.map(size=>(
-            <Label as="p" size={size} key={size+'kr'}>
+            <Links as="p" size={size} key={size+'kr'}>
               ({size}) 다람쥐 헌 쳇바퀴에 타고파
-            </Label>
+            </Links>
           ))
         }
         </section>
@@ -86,9 +86,9 @@ export const Size = {
         <section className='flex flex-col gap-4'>
         {
           sizes.map(size=>(
-            <Label as="p" size={size} key={size+'en'}>
+            <Links as="p" size={size} key={size+'en'}>
               ({size}) The quick brown fox jumps over the lazy dog
-            </Label>
+            </Links>
           ))
         }
         </section>
